lib/zodb/zstor_2zurl/NEO: support > 1 master nodes
The old code raised an explicit exception when converting a NEO storage
with > 1 master nodes into a URI. Perhaps the rationale for this exception
was that there isn't any agreed on order of master nodes in a NEO URI,
which means that building a URI from such a storage could potentially
break the invariant that any client which points to the same storage
should result in the same WCFS mountpoint.
With https://lab.nexedi.com/levin.zimmermann/wendelin.core/commit/6f5196fa10973473ea247c0d270e4ba932db1165 we can now rely on
WCFS mountpoint calculation to always return the same mountpoint even if
the order of master node addresses differ. Therefore we can drop this
exception and allow WCFS to support NEO clusters with more than one master.