lib/mntpt: Add explicit filtering depending on scheme
In the old source code we already filtered NEO URI by dropping
credentials, but we applied this filter to any URI, not only the NEO
one. This patch adds a mechanism to apply various filter according to
the specific storage type. Starting with this new patch, 'zstor_2mntpt'
also refuses to calculate the mountpoint for URI with unknown scheme.