qa: lib/tests += test_mntpt
After moving WCFS mountpoint calculation to a dedicated module, we can
now test this module for correctness. It's important that different
clients which point to the same storage always result in the same
mountpoint, even if their used zodburi slightly differs (e.g. due to
different client-side parameters or different paths of encryption).