qa: lib/tests/zodb += zurl_normalize_main
After moving zurl filtering to a dedicated function, we can
now test this function for correctness. It's important that different
clients which point to the same storage always result in the same
zodburi, even if their initial user-specified zodburi slightly differs
(e.g. due to different client-side parameters or different paths of encryption).

/reviewed-by @kirr
/reviewed-on https://lab.nexedi.com/nexedi/wendelin.core/merge_requests/17