lib/tests: Fix flaky zstor_2zurl test
Kirill noted that https://lab.nexedi.com/nexedi/wendelin.core/commit/fb62030103c781390385ce2346e435ccc5e6a303 introduced a regression [1]:
'test_zstor_2zurl' sometimes passes and sometimes fails. The reason for
this is that there is no deterministic order of master nodes in
'NodeManager.getMasterList()', which is why there is no specified
order of master node addresses in a zurl [2]. We don't want to normalize
a zurl returned by 'zstor_2zurl' as we need some of the client-specific
parameters as SSL file paths, so we rather fix the test to allow any
possible order of NEO master nodes in the zurl.

[1] https://lab.nexedi.com/nexedi/wendelin.core/merge_requests/17#note_188102
[2] https://lab.nexedi.com/nexedi/wendelin.core/blob/fb620301/lib/zodb.py#L414