wcfs: Add caller to WatchLink
A watch link is a link between WCFS and a specific client. Before this
patch, watch links didn't memorize basic information of the client (pid
and owner), now this information is added. This can be practical in case
the server needs to send signals to the client process. It's needed in
order to implement protection against faulty clients [1].

[1] https://lab.nexedi.com/nexedi/wendelin.core/blob/38dde766/wcfs/wcfs.go#L186-208