Fix `wendelin_cy_searh_in_dirs`.
The interface of the function `search_include_directories` has changed
in recent versions of Cython.
This updates the replacement used by wendelin so that it works for
both newer and older versions.

Note that wendelin-core still does not work with Cython >= 3, as that
version refuses to compile Python functions that can throw C++
exceptions (apparently, mixing C++ exceptions and Cython-generated
code is not considered safe).