Iterate the example data as a string.
The example data ('abcdefghij') should be iterated as a string, and not
as `bytes`.
Iterating a `bytes` object in Python 3 returns `int`, not `bytes`,
giving the following error on decoding:

```
--- FAIL: TestΔBTail (1.26s)
panic: root['treegen/values']: key ['b']: expected str, got int64 [recovered]
        panic: file:///tmp/TestΔBTail2213686135/001/1.fs: @03f97a31f270e722: get blktab: root['treegen/values']: key ['b']: expected str, got int64 [recovered]
        panic: file:///tmp/TestΔBTail2213686135/001/1.fs: @03f97a31f270e722: get blktab: root['treegen/values']: key ['b']: expected str, got int64
```

Instead, we can iterate as a string and encode each character to `bytes`
afterwards.