wcfs: Add tests to show wcfs kills dead clients
When a WCFS client doesn't respond in time to a pin request by
WCFS, the server attempts to kill this client [1]. However, there is
a possibility that after a server sends a pin request, a client stops
due to other reasons (for instance: restarted by other program). Then
the server shouldn't attempt to kill the client as this emits false
alarm logs about a faulty client whereas client isn't faulty but just
restarted at pin time.

[1] See https://lab.nexedi.com/nexedi/wendelin.core/-/commit/c559ec1a9