wcfs: Add tests to show WCFS kills dead or dying clients
When a WCFS client doesn't respond to a pin request in time, the server
attempts to kill it [1]. However, there are cases where the client may
stop for unrelated reasons (e.g. being restarted by another program)
after the pin request is sent. In such situations, WCFS should not
forcefully kill the client, as this leads to misleading logs suggesting
the client was faulty, when in fact it was simply restarted.

Also, if WCFS is faster than the client's graceful shutdown process,
force-killing the client can interfere with that shutdown.

This commit adds tests that reproduce these scenarios and verify
that WCFS only kills clients when truly necessary.

Note: these tests currently fail, as WCFS still kills dead or dying clients.

[1] https://lab.nexedi.com/nexedi/wendelin.core/-/commit/c559ec1a9