wcfs: py: Log with date and time present
So that we can correlate py log with go log outputs which is important
to analyze things in case of a trouble.

Standard py logging has support only for millisecond-level precision
(.msecs attribute of log record), but millisecond is not enough a
precision to reconstruct flow of FUSE events of what was going on. So we
need to increase this to microsecond precision manually and to match
timestamps precisions emitted by wcfs.go .

For consistency with wcfs.go also adjust wcfs.py log format to match
that of wcfs.go because both those py and log levels constitute one
program and log simultaneously.

Before this patch wcfs.py log output was like this:

    WARNING:wcfs:fuse_unmount /dev/shm/wcfs/a027435914736c5128c7f5c4931a7a59567a9e86: ...

After the patch it looks like this:

    W0421 14:16:11.101631 3857246 __init__.py:582] wcfs: fuse_unmount /dev/shm/wcfs/a027435914736c5128c7f5c4931a7a59567a9e86: ...

/reviewed-by @levin.zimmermann
/reviewed-on https://lab.nexedi.com/nexedi/wendelin.core/-/merge_requests/36