wcfs: py: Switch loglevel from WARNING -> INFO for wcfs.py commands
As it turned out managing and monitoring wcfs server turned out to be
not very robust. For example if wcfs.go is dadlocked, then `wcfs.py
status` will also hang and things like that.

From this point of view for better tracibility it makes more sense to
see the informative messages that wcfs.py commands are trying to emit.

/reviewed-by @levin.zimmermann
/reviewed-on https://lab.nexedi.com/nexedi/wendelin.core/-/merge_requests/36