bigarray/zodb: Switch from inspect.getargspec -> inspect.getfullargspec
--------
kirr:

In py3.11 inspect.getargspec was removed

    bigarray/array_zodb.py:40: in <module>
        _ = inspect.getargspec(BigArray.__init__)
    E   AttributeError: module 'inspect' has no attribute 'getargspec'

and https://docs.python.org/3/whatsnew/3.11.html says to replace it with
inspect.getfullargspec. In fact inspect.getargspec was deprecated since
py3, so we can do the change not only for py3.11+, but for any py3
version.

For our use-case getfullargspec seems to be drop-in replacement for
getargspec, so we should be ok doing that.